Add files via upload
Intro to Nets 2021 Hackathon Assignment - Quick Maths ⓶+⓶-①=③👑 Version 1.0

Write a client-server application which will implement a fast-paced math game. The players in this game get a random math question and are supposed to answer it correctly as fast as possible. We implemented both a client application and a server application.